+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## [1.0.3] - 03/09/2020
+
+* Added pushAndRemoveUntil and updated pushNamedAndRemoveUntil function.
+
 ## [1.0.2] - 06/05/2020
 
-* Fixed error on goBack function
+* Fixed error on goBack function.
 
 ## [1.0.0] - 19/03/2020
 
-* Updated to 1.0.0
+* Updated to 1.0.0.
 
 ## [0.0.3] - 26/02/2020.